🏷️ Use the same audio typing for REDmods in LO and installers
