+++ src/load_order.types.ts
@@ -1,3 +1,4 @@
+import { REDmodCustomSound } from "./installers.types";
 import {
   VortexWrappedDeserializeFunc,
   VortexWrappedSerializeFunc,
@@ -10,18 +11,10 @@ export interface LoadOrderer {
   serializeLoadOrder: VortexWrappedSerializeFunc;
 }
 
-export interface REDmodEntry {
+export interface LoadOrderEntryREDmod {
   folder: string;
   enabled: boolean;
   deployed: boolean;
   deployedVersion: string;
-  customSounds?: Array<REDSound>;
-}
-
-export interface REDSound {
-  name: string,
-  type: string,
-  file: string,
-  gain: number,
-  pitch: number
+  customSounds: REDmodCustomSound[];
 }
